# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get subjects list",
                "operationId": "7ec04c0df2c6d2830d25036d6f4d4b61",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subjects": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Group id",
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "description": "Subject name",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Creating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "updated_at": {
                                                        "description": "Updating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Store subject",
                "operationId": "f612fe6df780898b985b37bd4ed82da3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Subject name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subjects/{subject}": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Show subject",
                "operationId": "8d258e774868d03016b16a8f4c354e04",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Subject id",
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "description": "Subject name",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Creating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "updated_at": {
                                                        "description": "Updating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Update subject",
                "operationId": "93b54cc3883b8b943b24c908891dbb34",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Subject name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Delete subject",
                "operationId": "81e875ba9c5c29e8285393505250ffc4",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "Subject id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}